Refactor Json to support
