+++ web/router.ex
@@ -2,6 +2,7 @@ defmodule Exq.RouterPlug do
   require Logger
   require EEx
   alias Exq.RouterPlug.Router
+  alias Exq.Support.Json
 
   def init(options) do
     if options[:exqopts] do
@@ -54,7 +55,7 @@ defmodule Exq.RouterPlug do
       end
       qtotal = "#{Enum.sum(queue_sizes)}"
 
-      {:ok, json} = Poison.encode(%{stat: %{id: "all", processed: processed, failed: failed, busy: busy, scheduled: scheduled, enqueued: qtotal}})
+      {:ok, json} = Json.encode(%{stat: %{id: "all", processed: processed, failed: failed, busy: busy, scheduled: scheduled, enqueued: qtotal}})
       conn |> send_resp(200, json) |> halt
     end
 
@@ -70,18 +71,18 @@ defmodule Exq.RouterPlug do
       end
       all = %{realtimes: f ++ s}
 
-      {:ok, json} = Poison.encode(all)
+      {:ok, json} = Json.encode(all)
       conn |> send_resp(200, json) |> halt
     end
 
     get "/api/failures" do
       {:ok, failed} = Exq.Api.failed(conn.assigns[:exq_name])
       failures = for f <- failed do
-        {:ok, fail} = Poison.decode(f, %{})
+        {:ok, fail} = Json.decode(f, %{})
         Map.put(fail, :id, fail["jid"])
       end
 
-      {:ok, json} = Poison.encode(%{failures: failures})
+      {:ok, json} = Json.encode(%{failures: failures})
       conn |> send_resp(200, json) |> halt
     end
 
@@ -103,8 +104,8 @@ defmodule Exq.RouterPlug do
       {:ok, processes} = Exq.Api.processes(conn.assigns[:exq_name])
 
       process_jobs = for p <- processes do
-        {:ok, process} = Poison.decode(p, %{})
-        {:ok, pjob} = Poison.decode(process["job"], %{})
+        {:ok, process} = Json.decode(p, %{})
+        {:ok, pjob} = Json.decode(process["job"], %{})
         process = Map.delete(process, "job")
         process = Map.put(process, :job_id, pjob["jid"])
         process = Map.put(process, :id, "#{process["host"]}:#{process["pid"]}")
@@ -115,25 +116,25 @@ defmodule Exq.RouterPlug do
       processes = for [process, _job] <- process_jobs, do: process
       jobs = for [_process, job] <- process_jobs, do: job
 
-      {:ok, json} = Poison.encode(%{processes: processes, jobs: jobs})
+      {:ok, json} = Json.encode(%{processes: processes, jobs: jobs})
       conn |> send_resp(200, json) |> halt
     end
 
     get "/api/queues" do
       {:ok, queues} = Exq.Api.queue_size(conn.assigns[:exq_name])
       job_counts = for {q, size} <- queues, do: %{id: q, size: size}
-      {:ok, json} = Poison.encode(%{queues: job_counts})
+      {:ok, json} = Json.encode(%{queues: job_counts})
       conn |> send_resp(200, json) |> halt
     end
 
     get "/api/queues/:id" do
       {:ok, jobs} = Exq.Api.jobs(conn.assigns[:exq_name], id)
       jobs_structs = for j <- jobs do
-        {:ok, job} = Poison.decode(j, %{})
+        {:ok, job} = Json.decode(j, %{})
         Map.put(job, :id, job["jid"])
       end
       job_ids = for j <- jobs_structs, do: j[:id]
-      {:ok, json} = Poison.encode(%{queue: %{id: id, job_ids: job_ids, partial: false}, jobs: jobs_structs})
+      {:ok, json} = Json.encode(%{queue: %{id: id, job_ids: job_ids, partial: false}, jobs: jobs_structs})
       conn |> send_resp(200, json) |> halt
     end
